Refactor Course page
