+++ src/pages/course/index.js
@@ -1,0 +1,12 @@
+import CourseIcon from "@mui/icons-material/List";
+
+import CourseList from "./CourseList";
+import CourseCreate from "./CourseCreate";
+import CourseEdit from "./CourseEdit";
+
+export default {
+    create: CourseCreate,
+    edit: CourseEdit,
+    icon: CourseIcon,
+    list: CourseList
+};
